Renamed codes.py to utils.py
