+++ django_eth_events/tests/test_event_listener_exec.py
@@ -6,10 +6,10 @@ from django_eth_events.factories import DaemonFactory
 from django_eth_events.event_listener import EventListener
 from django_eth_events.web3_service import Web3Service
 from django_eth_events.models import Daemon, Block
-from django_eth_events.tests.codes import CentralizedOracle
+from django_eth_events.tests.utils import CentralizedOracle
 from web3 import TestRPCProvider
 from json import loads
-from django_eth_events.tests.codes import centralized_oracle_abi, centralized_oracle_bytecode
+from django_eth_events.tests.utils import centralized_oracle_abi, centralized_oracle_bytecode
 from django_eth_events.utils import remove_0x_head
 
 
@@ -34,7 +34,7 @@ class TestDaemonExec(TestCase):
             {
                 'NAME': 'Centralized Oracle Factory',
                 'EVENT_ABI': centralized_oracle_abi,
-                'EVENT_DATA_RECEIVER': 'django_eth_events.tests.codes.CentralizedOraclesReceiver',
+                'EVENT_DATA_RECEIVER': 'django_eth_events.tests.utils.CentralizedOraclesReceiver',
                 'ADDRESSES': [self.centralized_oracle_factory_address[2::]]
             }
         ]
@@ -47,7 +47,7 @@ class TestDaemonExec(TestCase):
         self.provider = None
 
     def test_create_centralized_oracle(self):
-        self.assertEqual(CentralizedOracle().len(), 0)
+        self.assertEqual(CentralizedOracle().length(), 0)
         self.assertEqual(0, Daemon.get_solo().block_number)
         self.assertEqual(0, Block.objects.all().count())
 
@@ -55,7 +55,7 @@ class TestDaemonExec(TestCase):
         tx_hash = self.centralized_oracle_factory.transact(self.tx_data).createCentralizedOracle('QmYwAPJzv5CZsnA625s3Xf2nemtYgPpHdWEz79ojWnPbdG')
         self.assertIsNotNone(tx_hash)
         self.listener_under_test.execute()
-        self.assertEqual(CentralizedOracle().len(), 1)
+        self.assertEqual(CentralizedOracle().length(), 1)
         self.assertEqual(1, Daemon.get_solo().block_number)
 
         # Check backup
@@ -68,7 +68,7 @@ class TestDaemonExec(TestCase):
         accounts = self.web3.eth.accounts
         self.web3.eth.sendTransaction({'from': accounts[0], 'to': accounts[1], 'value': 5000000})
         self.assertEqual(0, Block.objects.all().count())
-        self.assertEqual(CentralizedOracle().len(), 0)
+        self.assertEqual(CentralizedOracle().length(), 0)
         self.assertEqual(1, self.web3.eth.blockNumber)
 
         # Create centralized oracle
@@ -76,7 +76,7 @@ class TestDaemonExec(TestCase):
             'QmYwAPJzv5CZsnA625s3Xf2nemtYgPpHdWEz79ojWnPbdG')
         self.assertIsNotNone(tx_hash)
         self.listener_under_test.execute()
-        self.assertEqual(CentralizedOracle().len(), 1)
+        self.assertEqual(CentralizedOracle().length(), 1)
         self.assertEqual(2, Daemon.get_solo().block_number)
         self.assertEqual(2, Block.objects.all().count())
         self.assertEqual(2, self.web3.eth.blockNumber)
@@ -99,6 +99,6 @@ class TestDaemonExec(TestCase):
         Block.objects.filter(block_number=1).update(block_hash=block_hash)
 
         self.listener_under_test.execute()
-        self.assertEqual(CentralizedOracle().len(), 0)
+        self.assertEqual(CentralizedOracle().length(), 0)
         self.assertEqual(2, Daemon.get_solo().block_number)
         self.assertEqual(2, Block.objects.all().count())
